+++ src/JsonApi/Exception/RequestBodyInvalidJson.php
@@ -38,7 +38,7 @@ class RequestBodyInvalidJson extends JsonApiException
         $errorDocument = new ErrorDocument();
 
         if ($this->includeOriginalBody) {
-            $errorDocument->setMeta(["original" => json_decode($this->request->getBody(), true)]);
+            $errorDocument->setMeta(["original" => $this->request->getBody()->__toString()]);
         }
 
         return $errorDocument;